svc-flask/requirements.txt
Flask_Script==2.0.6
Flask==1.0.2
requests==2.22.0
Flask_Cors==3.0.7
Flask_RESTful==0.3.7
pytest==4.4.1
coloredlogs==10.0
python-dotenv==0.10.3
netaddr==0.7.19
python-Levenshtein==0.12.2